+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 
 *Release Date: *
 
+- Added option when converting Candle to include the readings
 - Fixed
     - Fixed Minor bug in _find_calc_index